refactor(workspace): extract teamclaw-gateway crate from monolithic src-tauri
Move gateway implementation files (discord, email, feishu, kook, wechat, wecom),
config types, session management, i18n, and shared helpers into a separate
workspace crate. Heavy deps (serenity, imap, lettre, tokio-tungstenite, image,
yup-oauth2) now live in the new crate. Tauri command wrappers remain in the
main crate's gateway/mod.rs with pub use re-exports for backward compatibility.
