+++ src-tauri/Cargo.toml
@@ -40,15 +40,9 @@ tokio = { version = "1", features = ["rt-multi-thread", "macros", "time", "sync"
 reqwest = { version = "0.11", features = ["json", "blocking", "stream", "rustls-tls-native-roots"], default-features = false }
 notify = { version = "8", features = ["macos_fsevent"] }
 notify-debouncer-mini = "0.7"
-serenity = { version = "0.12", default-features = false, features = ["client", "gateway", "model", "rustls_backend"] }
 futures = "0.3"
 base64 = "0.22"
 tokio-tungstenite = { version = "0.29", features = ["native-tls"] }
-imap = "2"
-native-tls = "0.2"
-mail-parser = "0.9"
-lettre = { version = "0.11", features = ["native-tls", "smtp-transport"] }
-yup-oauth2 = "12"
 flate2 = "1"
 tar = "0.4"
 minisign-verify = "0.2"
@@ -66,6 +60,9 @@ nanoid = "0.4"
 markitdown = "0.1.10"
 # STT crate
 teamclaw-stt = { path = "crates/teamclaw-stt" }
+
+# Gateway crate
+teamclaw-gateway = { path = "crates/teamclaw-gateway" }
 
 # RAG crate
 teamclaw-rag = { path = "crates/teamclaw-rag" }
@@ -124,7 +121,6 @@ hkdf = "0.12"
 hex = "0.4"
 getrandom = "0.2"
 url = "2"
-image = { version = "0.25", default-features = false, features = ["jpeg", "png", "gif", "webp"] }
 log = "0.4"
 
 # Global keyboard event monitoring (double-tap Ctrl for spotlight)
