+++ src-tauri/crates/teamclaw-gateway/src/session_queue.rs
